+++ ShoppingProject/ShoppingMvc/Models/Cards/Product.cs
@@ -9,9 +9,9 @@ namespace ShoppingMvc.Models.Cards
     {
         public string? Description { get; set; }
         [Column(TypeName = "smallmoney")]
-        public decimal SellPrice { get; set; }
+        public float SellPrice { get; set; }
         [Column(TypeName = "smallmoney")]
-        public decimal CostPrice { get; set; }
+        public float CostPrice { get; set; }
         public int CategoryId { get; set; }
         public Category? Category { get; set; }
         public string? ImageUrl { get; set; }
